Create README.mdAI ALGORİTMASI
https://1drv.ms/w/s!AuZP6u40PadPcnTOF7YKG9onJNI?e=8OphXh
FAIROB (VTR) PLAN 

1- FAIROB (forex artificial intelligence robot) works on MetaTrader4, Metatrader5, TradingView, cTrader,İnvesting, İnvesting Pro platforms. 

2-FAIROB works. 

2.1- FAIROB lot buying rules 

2.2- FAIROB lot selling rules 

3-FAIROB workflow document. 

4-User interaction rules with FAIROB. 

5-About FAIROB updates 

6-About the FAIROB update server(FUS). 

7-About FAIROB security requirements. 

 

 

1-FAIROB will work with VIOP with information. 

 

VIOP is the market where the shares of the biggest companies operating in Turkey and dollar, euro and gold can be bought and sold. VIOP products are offered for sale in varieties with three-monthly different pricing. For example, the January share of a company is 7 TL, the February share is 7.3, the March share is 7.7. VIOP starts trading at 9.30 am on weekdays and lasts until 6 pm in the evening, trading continues until 18.15 for gold, euro, dollar. It is a very fast-moving market, with more than one million trading orders per day. Generally, trades take as 1 unit, 10 units, 25 units. Very rarely, 200 and 500 buy or sell transactions take place. Price update in VIOP products is a common situation due to the inflationary effect. The cheapest lot has a price range from 30 TL to 2000 TL. Stocks are grouped as industry, banking, aviation, gold production companies, iron and steel production companies, holdings, and rises and falls are usually grouped. The market is very sensitive to the influence of the state and often operates under the influence of intervention. State intervention is mostly seen in dollar, euro, gold pricing, it is not felt at all in stocks. The actors that have an impact on the stocks are foreign states in the Asian continent, with the most influence coming from Japan. Market ignorance caused by the time difference makes the market vulnerable. The manipulative information pollution is too much and it will surely succeed, the reverse transaction will result in the bankruptcy of the investor. At least once a month, a stock gains more than 10% in value per day and will definitely appreciate another 5% at the next day's opening. When a stock gains more than 10% on two consecutive days, it's very surprising, and it happens every 50 years. 

 

2.1- FAIROB Works 

The VTR is an earning robot and must take actions that are adaptive in nature. For this reason, VTR can gain gain over the increase in values. Losses are very likely to reduce the gain of the VTR. VTR buys upside stocks and gold, dollars, euros when their prices are on the rise and sells them when they reach a certain profit limit. When the VTR reaches the specified profit limit, it must notify the group of robots that do the same before selling; otherwise, it will cause damage to the robots that make purchases after it. Therefore, the last 4 robots with the same stock will exit the earliest, and the stock selling order will continue in the last 4 top 1 order. VTR follows the determined data, provides the necessary communication for other robots to follow correctly and allows other robots to follow. 

VTR is a trading robot with a cold-blooded nature. It detects alarming causes and warns other robots of the threat. For example, if the price increase in foreign currency is at a level that requires government intervention, it warns other robots and prevents them from making purchases. VTR purchase does not necessarily have to make a purchase as it does not feel obligatory, on the contrary, the sale must be carried out according to the determined obligations. VTR has to perform the market optimization function while making a profit. Determined profit rates reduce the risk of loss by ninety percent. VTR does not promise its user more than one hundred percent profit in 1 month; however, it can guarantee that no loss will occur. The main function of VTR is to keep the investor away from stress, to maximize human's limited tracking ability within himself and to make a profit without breaking the economic rules. 


Signed-off-by: ROBERT DA VİNCİ TIME MACHINE <info@johnsonssdlccompany.com>